refactor: move JSON fetcher files into its own dir
