+++ src/app/common/json/fetcher/json-fetcher.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core'
-import { CONTENTS_DIR } from '../directories'
+import { CONTENTS_DIR } from '../../directories'
 
 export const JSON_DATA_DIR = new InjectionToken<string>('JSON data dir', {
   factory: () => CONTENTS_DIR,
